Add table to define super characteristics
This table is used as the primary definition of individual super
characteristics.

Note that this commit is not the final design, but the original design I
think there is value in this design and want to show it in the as part
of the evolution of the product.  As this design sits, we'll implement
this in a view and put the validation rules into a normalized relational
structure... kind of a coin toss though as most of the time we'll be
making use of the validation rules in JSON form and it's a clearer
representation of a rules engine than is a relational model.  Any way,
in this case a coin toss is always defaulted to relational data so that
is what we'll do.