Add flag field flag and user visible flag
For the data_type table we add fields which allow a data type to be
designated as a boolean type and we also add a field for whether or not
a data type is user assignable via the UI.

For the validator_type table we add a type category for boolean types.
The user assignable flag already existed on this table.