Add file to add "early" musemetrics
These are values that may be referenced during the loading of seed data
so we need them fairly early in the process.  So long as creating the
metric itself doesn't depend on seed data we can populate here.

Note that the current settings of the auditing config should default to
being off on installation so as not to cause a surprising behavior.  For
development however, full auditing on is likely to be useful.