Remove "required" flag and add "searchable" flag
The idea of "adding" a characteristic to a table is not valid any longer
so we get rid of the field that tracked whether or not adding a field
was required.  However, we will try to "fake" existing characteristic
data so we need to know whether or not a characteristic is searchable.