Add trigger managing entity system lock flag
When there are no more managing packages for a system locked entity in
the entity_package table, remove the system locked restriction from the
entity in the entity table.  This effectively turns the entity into a
user managed record which would be consistent with the possible desire
to use the entity outside of the package context.