Refactor include ref; add entity cols & populate
We refactor the data script include from the originally planned general
data acces library to one specific to entities (there will be others).

We add columns to the entity related xtreewidget.

We add functions to clear the loaded data from the form.  We add a
function to populate them, starting with the entity related xtreewidget.