Remove "set"; add set buttons function
Because the setup form doesn't use the "set" function for its subforms,
we just remove that and do our initialization directly since we don't
need parameters.

We also add a function to manage the button state based on permissions
and whether or not a record is system locked.  We make some calls based
on that state, though that effort is still work in progress.