Entity editing support; improve data table naming
We want to allow changing display names pretty consistently.  We also
want to allow changing the id column if needed, but we can only allow
the id column change in cases where there isn't already data stored with
the previous column saved.  In this commit we add fuctionality to
achieve both goals.

Along the way it became clear we needed a better method to handle long
entity table names since we could overrun the id column name pretty
easily.  We added infrastructure to accomplish that here; we truncate
long names and simply return nothing if we can't meet a feasible length.