Change version to something more representative
I've had a version 0.1.0, but really we shouldn't have a 0.1.0 until we
release an alpha quality version of the extension and we're not there
yet.  So while we're at 0.0.x we'll create the second digit as
alpha/beta indicator with 0.1.x being alpha and 0.2.x being beta.  Once
we reach version 1.0.0 we'll follow something closer to semver.