Change group assignment lock lookup source
In the group/entity assignment lookups, we were considering if the
entity was locked.  That was wrong: the right consideration is if the
group/entity assignment itself is system locked.  We need to be careful
about being sure that locks deal directly with the record type in
question or that we're trying to satisfy.