Add check for record return and fix params
A single bug led to both changes.  Rather than passing "pParams" to the
query, we passed "{pParams: pParams}" this caused a failure.  The
failure was so far up the stack, it wasn't even obvious if we were
passing the parameters; so we added a data return check in the
getGroupById function... we should get at least get a record for any
call and we want to see the parameters in play.