Remove vestigial button/spacer; bump version
The prior commit while completely functional did leave a vestigial
button for assignment of groups to entities; we now just do that on
the actual group window so this button has no function other than adding
user complexity.

Also, the last commit we managed to not quite bump the version number
due to some undo that had to happen... <sigh>.  We bump here which is
just as well since this really is more accurately the group complete
milestone.