Simplify library api, remove blanket try/catch
The Muse Systems code practice in the library pPublicApi calls were to
wrap the non-public function calls in try/catch blocks.  I've never
been completely happy with this and on further reflection, I've decided
to change the practice.

If we accept that the library public api functions do three qualities:

1) check parameters which can throw exceptions on failure;
2) calls a single private function;
3) calls a private function that is only called by this public api
   function.

the utility of catching calls to the private function don't seem useful.
We will always know who called the private function (which should
try/catch any possibly failing logic), so the extra stack trace doesn't
seem useful, and we don't want to catch any exceptions we throw in the
public api function in the same function anyway. Assuming that points 1,
2, & 3 are rules for public api functions then the try/catch is
extraneous and should be avoided; if more complex logic, like calling a
succession of private functions or where multiple public api functions
call the same private function, then the stack trace we get from the
try/catch block can be useful in understanding how we arrived at the
private function.