Add database level validator violations checks
Since it is possible to create situations where a super characteristic
conditional validation rule could never be satisfied because a subject
characteristic would not be able to reach it's object characteristics,
we need to check if that is going to happen and prevent it.  We cannot
prevent all unresolvable validations, but this class of subject/object
visibility constraints we can check before we allow them to happen.

This commit adds the trigger functions and triggers which call the
previously defined queries.  We may still have an issue in the group
add to entity query (a group may bring its own dependencies which do
not exist for the entity prior to the add... that may still cause an
exception), but other than that possibility I think this is where we
want to be.