Add new generalized types; update seed data
We add generalized data types to support list of value and multiple
select list of value fields.  Originally we just assumed this was
one of the other more basic types, such as numeric, since we could
store the result that way.  However, by having flags for LOV types we
can code the handling a bit more now without having to do specific hard
coding later (or at least less of that).

We refactor the seed data to include the new columns and to make use
of the ON CONFLICT clause of INSERT.  Since we're PostgreSQL 9.5 or
later, we don't need to play select tricks like we were.