Big commit wish makes the Group Layout functionality complate
This includes form generation and preview ability.  Another major
refactor in regards to data storage; rather than expecting to save via
JSON, we now create and drop musesuperchar entity table columns for the
data points as needed.  Turns out the current design is really not
unstructured in the sense that sc/grp/entity combination allows for
"unassigned" characteristics for the entity... it may be null, but it
will always exist (or need to be presumed to exist)... so just bite the
bullet and add/drop the columns as needed.  We don't default/not null
the values in any case so that it performs reasonably well.

We may want to add a delete lock of some sort since an accidental
Group Layout Item delete is not hard to do and could very possibly
result in lost data.

We bump the version number as we're almost complete and this piece is
significant.