Fix extra empty LOV value on first entry
When a new LOV superchar is added and the LOV values are hand entered,
the first time an empty starting element would always show up.  The
reason is that we convert a PostgreSQL Array to a comma sepatated
string and then use the JavaScript split function to create a JS array.
When split is called on an empty string, you get an array with element 0
being an empty string.