Add UI form back script gen; data hook system
This is a large change that adds the UI form backing script generation
routines and templates.

We also introduce the hook system in the entity scripts.  Thie allows
external scripts (or our conditional validation rules) to be set on an
event hook basis.  The events supported are onNew, onLoad,
beforeSetValue, afterSetValue, beforeSave, afterSave.