Add SuperChar integration to xTuple functionality
This is a large commit which closes the loop on bringing the up-to-now
isolated functionality together in such a way that the Super Char
functionality can be easily brought to standard record/form types.  The
key to this is the museScLoader module.  Originally I had intended the
museScWidget library to serve this purpose, but there's some cruft there
from earlier thinking that needs to be deprecated and removed; until
that happens that libary is too heavy for what we should be willing to
accept in the running forms.   Also the load name is more appropriate
for how the design evolved from planning to implementation.

There are also a large number of bug fixes in this commit which were
surfaced in the testing of the loader module and the action of the
generated forms attached to the Item record.