Remove filecluster type; add metrics for std recs
Right now the file cluster doesn't function in the originally expected
way and it's likely it's not useful.  We'll just comment it out for now
and in the 1.0 release we'll likely remove it outright, unless there's
some revelation about it being useful.

We add some new metrics that will be referenced by standard record/forms
and will drive 1) the name of the parent container of the groups
tabwidget; 2) whether or not we show the xTuple out-of-box
characteristics tab;