Add Item form Super Char script; fix "new" bug
This marks the frst commit of a baseline script for integrating existing
forms/records with the Super Characteristic system.  This script is
meant to be generic enough to use with most record types as-is, but also
as an example of how to use the museScLoader library to start up a
Super Characteristic UI properly.

We also fix some save process bugs that would crop up when saving a new
record for the first time.  The form script couldn't handle the "new"
case where we'd start with a record designator of "new12345" and
transition to a proper database ID record designator.  This fixes the
sequence to allow that transition to happen.