Add a fix permissions file
When a user other than admin installs a package, the owner of the
created objects is the installing user rather than the more generic
admin.  While our explicitly created objects are correctly set to the
correct owner, the xTuple implicitly created objects (such as the
pkg* tables) are not.  This ensures they are owned by admin.