Add Lov Query Override functionality
This functionality allows a scripting client developer to add a
specialty query to populate a combobox with a specific query.  For
example, one use case involves making the list a list of item aliases
that varies with the contextually set item record.  We cannot define the
super char query with this normally since we don't have a source for the
item number.  With this change we can accomplish that with a script.