Version bump for production
We're now officially in production with a client and as such, semver
versioning says that we should by now be at version 1.0.0.

There are no breaking public API changes since the last published
version of 0.1.3, so in this case the version bump does not reflect
breaking API changes; just the move from 0.x.x to 1.x.x versions.