Add musextputils save hook framework
We make the sales order item script musextputils save framework aware.
This is because there can be multiple scripts which need to do work
before and after the native save function, but there are no hooks to
know which part of the process we're in.  Because of this, we have a
governing script in the musextputils package which now governs the
process and allows us to hook into the appropriate stage of the save for
our purposes.