Fix duplicating group tab issue
There was an issue where group tabs could be added to multiple times to
a loader based form, when that form was reused to display multiple
records (i.e. Sales Order Item Prev/Next button functionality).  Now we
check if we've already initialized the groups structure in any loader
based instance.

This issue has been observed in the past, but not logged as it has not
been commonly seen. it was found again testing for:
https://gitlab.com/musesystems-opensource/musesuperchar/issues/3