Add func to get ScWidget & current ScWidget dataId
In order to make default superchar forms (item, customer, etc.)
scriptable outside the musesuperchar package, we need to be able to get
more information from outside of the default form script.  In this case
we add a public API function to the default superchar form scripts which
allows other, later scripts to retrieve the SuperChar Widget Object
created by the loader and (before this) otherwise private to the default
form script.

We also add a public API function to the ScWidget created by the loader
to retrieve its current Data Rec. ID.  There are public functions which
rely on knowing the correct Data Rec. ID and this state is already kept
internally by the generated ScWidget object created by the loader.