Refactor numeric values to coalesce with 0
The various numeric widgets are based on the musextputils
numericLineEdit widget and that widget doesn't accept nulls.  Since we
don't automatically assume defaults on creation, we need to ensure that
we make some basic assumption when we construct the widgets.

In the long run, the musextputils should accept nulls and allow client
applications to decide whether or not that's a valid condition.  But for
now that's too big a breaking change and so we'll change our assumptions
here.

Closes #4