Fix missing parsedParams refs & refactor format
We rely on having simple values in configuration data so we started parsing
parameter objects which could arrive at a function as the expected simple values
or as those values wrapped in some sort of QVariant object wrapper.  However,
this was left out of this script; this script isn't fully built and so no real
rigorous testing has happened...

We inadvertantly auto-refactored code to remove trailing whitespace and meet
Prettier.io formatting convensions.  We need these and it will take too much
work to unwind so we'll commit those here.