Fix errors via Sales -> Customer -> Workbench...
There were errors thrown when users entered the customer form via the Sales ->
Customer -> Workbench... path.  The form would open in edit mode, but have a
-1 cust_id.  We update the form to respond to mode changes that can happen via
this method.  We should watch memory related issues and such as I think we can
end up adding data records as different customers are loaded via this method,
but in practice this shouldn't be an issue made manifest to the user.

Please note that there is a MuseUtils fix required for this as well which will
be in version 2.9.1.

Resolves https://gitlab.com/musesystems-rwm/support/issues/22