Refactor getValidatorValuesStructure w/return val
This is an effort to be a little more functional and a little less dependent on
side effects for getting stuff done.  This was a good example of how depending
on side effects made it harder to understand what was going on.  We were setting
some global state in getValidatorValuesStructure and then counting on that
global state to have been set properly.  Reading where
getValidatorValuesStructure was being called and understanding where its result
was used made the relationship unclear (in addValidatorWidget).

This commit doesn't stop the script global state from being stored, but it also
returns a reference to that state so that callers of the
getValidatorValuesStructure can use the function and its results in an easier to
use form.