Add min columns and is space conserved columns
This adds new configuration points to the group definition.

The first allows a user to set the minimum number of columns a group layout
should generate (#10).  The second moves the isLayoutSpaceConserved global
configuration to the group definition level so that individual groups can
make use of space conservation or not (#13).