Add is row expansion allowed column
This is another new group level setting.  When set, the sections on any row of a
group can expand to take the full group widget width, even if there is more
space than they need or should take.  This is the old behavior and the default
behavior here.

When false, we'll add spacers to the end of the row to push the sections
together; trading more attractive sections for perhaps some ugly end of row
space.