Remove last spots where internal names are edited
We get rid of the last bits of the *_internal_name edit capability.  Now
internal names are set in stone.  There are remaining issues where updating an
internal name to a user defined value on creation doesn't seem to work without
resetting it to the system defined default, but that's a different, less
important issue.

This will resolve #14