Add virtual scdef support to data lib generation
The previous version relied on extracting the data fields to be managed by the
JS data lib file directly from the table where the data was persisted.  This
worked well and was pretty clean until we added virtual scdefs.  In that case
the fields aren't in the entity data table, but still needed to be managed like
any other data field from the JS perspective.