Refactor item initialization and refresh sequences
Due to https://gitlab.com/musesystems-od/support/issues/69, we need to update
the way we interact with the item form on 1) initialization via "set" and,
2) responding to changes in the current item being presented to the user.

The utilities package had to change the way it interacted with the item
availability workbench which changed how we needed to work here.

Note that with this change, we require musextputils v4.x.x.