Add support for the muserfq extension
This commit adds support for the Muse Systems Request for Quote for xTuple ERP
extension.  It does not require that the module is installed; if the muserfq is
not installed, these scripts will never run, though they'll still be loaded.

Each of the three scripts (RFQ Head, RFQ Item, and RFQ Vendor Item) have been
tested and seem to perform as expected in this commit.