Fix unfavored uniqueness constraint
It dawned on me after I saved the 1.5.1 that I was applying the uniqueness in a
disfavored manner.  While you can create unique indexes in postgresql, it's
suppose to simply be an implementation detail; to be fair, it can be useful to
do that directly since you can create an index concurrently and you cannot
create a constraint concurrently (and an index is created for a unique
constraint anyway).

After the remembered there being something about this, a quick searche led me
to: https://preview.tinyurl.com/ycsh78ob (I shortened original link)
which had a nice overview.

While the previous method probably would have worked fine (not sure about the
ON CONFLICT bit, which is the reason for doing this), but this is correct which
is better.