Fix issue with superchar saves when none defined
There was a problem saving records where there were no superchar groups defined
for the record type.  In those cases the scWidget would (correctly) be null,
but we still connected to the signals and didn't check if there was work to do
in the SC save function.

We don't futz with the signal/slot connects (those get problematic) so we simply
check if an scWidget exists in the save action prior to saving.