Fix improper unique on scdef/scgrp assign table
This fix reverses the unique constraint added in cbad9956.  While a unique
constraint was useful for regular superchars, where they should only ever be in
a group once, it was a problem for repeatable chars such as empty space or
horizontal line which can appear many times.

We'll have some rework in other packages (like those trying to do an ON CONFLICT
UPDATE or INSERT), but that's addressable.

Addresses: https://gitlab.com/musesystems-od/support/issues/76