Refactor script load order
This changes the load order of the JavaScript files.  This is mostly for the
benefit of "xtp override" sciprts, such as salesOrder or transferOrderItem and
making these work reliably across xTuple extensions.  As such we expect xTuple
native scripts to live a order values of 0 - 5, musextputils to be at 6, and
we'd be next in the dependency chain at 8.  This means we can assume
musextputils scripts to run before the superchar scripts, and the superchar
scripts to run before higher level consumers of our functionality, such as
muserfq.