Add ability to set/unset char display only mode
This change adds an API allowing client scripts to programatically set a data
point to be display only or not.  This field will not override mode settings, so
a view mode form that is requested to make a field editable will leave the field
in view only; mode enable/disable is set at the group level which overrides the
field level.... should the mode change, the request will be effective.

Also note that we do not distinguish between initially display only fields and
initially editable fields.  At the API level right now it feels right to treat
the user setup of the superchar as a default rather than an absolute... this
could change, but for now this should do.