Fix bad null check in museScLoader, rm workaround
The test for a non-existent parent record ID was too strict, treating
other "not-defined" values, such as undefined, as though they were IDs.

This meant that the code would then try to look up that was rightfully
strict about valid IDs and error.

This commit fixes that bug and also removes a workaround in
accountNumber.js which was triggering the error via an undefined value.