Fix purchaseOrder "releasePr" mode initialization
The initialization sequence is foobar when the purchase order form is
loaded via releasing a purchase request.  In that case, we have to
treat the form as re-entrant since the initial load of the form doesn't
have the correct mode or record ID until after initialization in many
instances (particularly if there's an existing purchase order).  In the
cases where the ID comes later, we don't have a full initialization
sequence and so we have to capture the scenario and force initialization
of the scWidget object.